requirements.txt
cowsay==6.1
google-cloud-storage>=2.14.0
requests>=2.28.0
pandas>=2.0.0
pyarrow>=14.0.0
